+++ src/Configuration.php
@@ -492,6 +492,16 @@ class Configuration implements
     protected $filename;
 
     /**
+     * The log file that has to be watched a import logging.
+     *
+     * @var string
+     * @Type("string")
+     * @SerializedName("log-file")
+     * @Accessor(setter="setLogFile", getter="getLogFile")
+     */
+    protected $logFile;
+
+    /**
      * Lifecycle callback that will be invoked after deserialization.
      *
      * @return void
@@ -1602,5 +1612,22 @@ class Configuration implements
     public function getFilename()
     {
         return $this->filename;
+    }
+
+    /**
+     * @return string
+     */
+    public function getLogFile()
+    {
+        return $this->logFile;
+    }
+
+    /**
+     * @param string $logFile The log file to use
+     * @return void
+     */
+    public function setLogFile($logFile)
+    {
+        $this->logFile = $logFile;
     }
 }