+++ CHANGELOG.md
@@ -6,12 +6,14 @@
 * Fixed #PAC-280: Extension library requires `etc` configuration directory
 * Fixed #PAC-206: Prevent finder mappings of different libraries to be overwritten
 * Fixed #PAC-318: Serialization exception when using a custom database configuration
+* Add #PAC-486: Add Getter and Setter for `--log-file` commandline parameter 
 
 ## Features
 
 * Add #PAC-89: Add debug email command + DebugSendPlugin
 * Add #PAC-57: Deleting dedicated attribute values via import by setting a configurable value
 * Add #PAC-349: Support for JMS serializer v2.0 + v3.0
+* Configuration for import explizit filename
 
 # Version 16.1.0